Fixed some checkstyle errors in gui module.
