+++ PL2/PL2-gui/src/main/java/nl/tudelft/pl2016gr2/gui/model/IPhylogeneticTreeNode.java
@@ -8,48 +8,48 @@ package nl.tudelft.pl2016gr2.gui.model;
  */
 public interface IPhylogeneticTreeNode {
 
-    /**
-     * Check if this node has a parent.
-     *
-     * @return if this node has a parent.
-     */
-    public boolean hasParent();
+  /**
+   * Check if this node has a parent.
+   *
+   * @return if this node has a parent.
+   */
+  public boolean hasParent();
 
-    /**
-     * Get the parent of this node.
-     *
-     * @return the parent of this node.
-     */
-    public IPhylogeneticTreeNode getParent();
+  /**
+   * Get the parent of this node.
+   *
+   * @return the parent of this node.
+   */
+  public IPhylogeneticTreeNode getParent();
 
-    /**
-     * Get the amount of direct child nodes (i.e. child nodes with an edge to this node).
-     *
-     * @return theamount of direct child nodes.
-     */
-    public int getDirectChildCount();
+  /**
+   * Get the amount of direct child nodes (i.e. child nodes with an edge to this node).
+   *
+   * @return theamount of direct child nodes.
+   */
+  public int getDirectChildCount();
 
-    /**
-     * Get the total amount of child nodes. This includes the direct child nodes and the indirect
-     * child nodes.
-     *
-     * @return the total amount of child nodes.
-     */
-    public int getChildCount();
+  /**
+   * Get the total amount of child nodes. This includes the direct child nodes and the indirect
+   * child nodes.
+   *
+   * @return the total amount of child nodes.
+   */
+  public int getChildCount();
 
-    /**
-     * Get the child at the given index.
-     *
-     * @param index the index of the child node.
-     * @return the child at the given index.
-     */
-    public IPhylogeneticTreeNode getChild(int index);
+  /**
+   * Get the child at the given index.
+   *
+   * @param index the index of the child node.
+   * @return the child at the given index.
+   */
+  public IPhylogeneticTreeNode getChild(int index);
 
-    /**
-     * Get the index of a direct child node.
-     *
-     * @param child a direct child node.
-     * @return the index of the child ndoe.
-     */
-    public int getChildIndex(IPhylogeneticTreeNode child);
+  /**
+   * Get the index of a direct child node.
+   *
+   * @param child a direct child node.
+   * @return the index of the child ndoe.
+   */
+  public int getChildIndex(IPhylogeneticTreeNode child);
 }
