+++ PL2/PL2-gui/src/main/java/nl/tudelft/pl2016gr2/gui/view/tree/GraphArea.java
@@ -11,79 +11,84 @@ package nl.tudelft.pl2016gr2.gui.view.tree;
  */
 public class GraphArea {
 
-    protected final double startX, endX, startY, endY;
+  protected final double startX, endX, startY, endY;
 
-    /**
-     * Create a graph area.
-     *
-     * @param startX the x coordinate of the left border of the graph area.
-     * @param endX the x coordinate of the right border of the graph area.
-     * @param startY the y coordinate of the top border of the graph area.
-     * @param endY the y coordinate of the bottom border of the graph area.
-     */
-    public GraphArea(double startX, double endX, double startY, double endY) {
-        this.startX = startX;
-        this.endX = endX;
-        this.startY = startY;
-        this.endY = endY;
-    }
+  /**
+   * Create a graph area.
+   *
+   * @param startX the x coordinate of the left border of the graph area.
+   * @param endX the x coordinate of the right border of the graph area.
+   * @param startY the y coordinate of the top border of the graph area.
+   * @param endY the y coordinate of the bottom border of the graph area.
+   */
+  public GraphArea(double startX, double endX, double startY, double endY) {
+    this.startX = startX;
+    this.endX = endX;
+    this.startY = startY;
+    this.endY = endY;
+  }
 
-	public double getStartX() {
-		return startX;
-	}
+  public double getStartX() {
+    return startX;
+  }
 
-	public double getEndX() {
-		return endX;
-	}
+  public double getEndX() {
+    return endX;
+  }
 
-	public double getStartY() {
-		return startY;
-	}
+  public double getStartY() {
+    return startY;
+  }
 
-	public double getEndY() {
-		return endY;
-	}
+  public double getEndY() {
+    return endY;
+  }
 
-    /**
-     * Get the width of the graph area.
-     * @return the width of the graph area.
-     */
-    public double getWidth() {
-        return endX - startX;
-    }
+  /**
+   * Get the width of the graph area.
+   *
+   * @return the width of the graph area.
+   */
+  public double getWidth() {
+    return endX - startX;
+  }
 
-    /**
-     * Get the height of the graph area.
-     * @return the height of the graph area.
-     */
-    public double getHeight() {
-        return endY - startY;
-    }
+  /**
+   * Get the height of the graph area.
+   *
+   * @return the height of the graph area.
+   */
+  public double getHeight() {
+    return endY - startY;
+  }
 
-    /**
-     * Get the center x coordinate of the graph area.
-     * @return the center x coordinate of the graph area.
-     */
-    public double getCenterX() {
-        return (endX - startX) / 2.0 + startX;
-    }
+  /**
+   * Get the center x coordinate of the graph area.
+   *
+   * @return the center x coordinate of the graph area.
+   */
+  public double getCenterX() {
+    return (endX - startX) / 2.0 + startX;
+  }
 
-    /**
-     * Get the center y coordinate of the graph area.
-     * @return the center y coordinate of the graph area.
-     */
-    public double getCenterY() {
-        return (endY - startY) / 2.0 + startY;
-    }
+  /**
+   * Get the center y coordinate of the graph area.
+   *
+   * @return the center y coordinate of the graph area.
+   */
+  public double getCenterY() {
+    return (endY - startY) / 2.0 + startY;
+  }
 
-    /**
-     * Check if the given coordinates are within the graph area.
-     * @param xCoord the x coordinate.
-     * @param yCoord the y coordinate.
-     * @return if the given coordinates are within the graph area.
-     */
-    public boolean contains(double xCoord, double yCoord) {
-        return xCoord >= startX && xCoord <= endX
-                && yCoord >= startY && yCoord <= endY;
-    }
+  /**
+   * Check if the given coordinates are within the graph area.
+   *
+   * @param xCoord the x coordinate.
+   * @param yCoord the y coordinate.
+   * @return if the given coordinates are within the graph area.
+   */
+  public boolean contains(double xCoord, double yCoord) {
+    return xCoord >= startX && xCoord <= endX
+            && yCoord >= startY && yCoord <= endY;
+  }
 }
