+++ PL2/PL2-gui/src/main/java/nl/tudelft/pl2016gr2/gui/view/tree/ViewNode.java
@@ -18,231 +18,231 @@ import nl.tudelft.pl2016gr2.gui.model.IPhylogeneticTreeNode;
  */
 public class ViewNode extends Circle {
 
-    private final static double NODE_RADIUS = 10.0;
-    private final static double NODE_DIAMETER = NODE_RADIUS * 2.0;
-    private final static Duration ANIMATION_DURATION = Duration.millis(750.0);
+  private final static double NODE_RADIUS = 10.0;
+  private final static double NODE_DIAMETER = NODE_RADIUS * 2.0;
+  private final static Duration ANIMATION_DURATION = Duration.millis(750.0);
 
-    private final IPhylogeneticTreeNode dataNode;
-    private final ViewNode parent;
-    private final ArrayList<ViewNode> children = new ArrayList<>();
-    private Line edge, elipsis;
-    private final TreeController controller;
-    private GraphArea graphArea;
-    private boolean isLeaf = true;
+  private final IPhylogeneticTreeNode dataNode;
+  private final ViewNode parent;
+  private final ArrayList<ViewNode> children = new ArrayList<>();
+  private Line edge, elipsis;
+  private final TreeController controller;
+  private GraphArea graphArea;
+  private boolean isLeaf = true;
 
-    /**
-     * Create a nl.tudelft.pl2016gr2.gui.view node.
-     *
-     * @param dataNode the data of the node.
-     * @param parent the parent nl.tudelft.pl2016gr2.gui.view node.
-     * @param graphArea the graph area in which the node has to be drawn.
-     * @param controller the controller of the tree.
-     */
-    private ViewNode(IPhylogeneticTreeNode dataNode, ViewNode parent, GraphArea graphArea,
-            TreeController controller) {
-        super(NODE_RADIUS);
-        this.dataNode = dataNode;
-        this.graphArea = graphArea;
-        this.parent = parent;
-        this.controller = controller;
+  /**
+   * Create a nl.tudelft.pl2016gr2.gui.view node.
+   *
+   * @param dataNode the data of the node.
+   * @param parent the parent nl.tudelft.pl2016gr2.gui.view node.
+   * @param graphArea the graph area in which the node has to be drawn.
+   * @param controller the controller of the tree.
+   */
+  private ViewNode(IPhylogeneticTreeNode dataNode, ViewNode parent, GraphArea graphArea,
+          TreeController controller) {
+    super(NODE_RADIUS);
+    this.dataNode = dataNode;
+    this.graphArea = graphArea;
+    this.parent = parent;
+    this.controller = controller;
 
-        double red = Math.abs(((dataNode.getChildCount() * 13) % 200) / 255d);
-        double green = Math.abs(((dataNode.getChildCount() * 34) % 200) / 255d);
-        double blue = Math.abs(((dataNode.getChildCount() * 88) % 200) / 255d);
-        this.setFill(new Color(red, green, blue, 1.0));
+    double red = Math.abs(((dataNode.getChildCount() * 13) % 200) / 255d);
+    double green = Math.abs(((dataNode.getChildCount() * 34) % 200) / 255d);
+    double blue = Math.abs(((dataNode.getChildCount() * 88) % 200) / 255d);
+    this.setFill(new Color(red, green, blue, 1.0));
 
-        this.setCenterX(this.getRadius() + graphArea.startX);
-        this.setCenterY(graphArea.getCenterY());
+    this.setCenterX(this.getRadius() + graphArea.startX);
+    this.setCenterY(graphArea.getCenterY());
+  }
+
+  /**
+   * Get the data of this node.
+   *
+   * @return the data of this node.
+   */
+  public IPhylogeneticTreeNode getDataNode() {
+    return dataNode;
+  }
+
+  /**
+   * Draw a root node and all of its children which fit on the screen.
+   *
+   * @param root the root node.
+   * @param controller the controller of the tree.
+   * @return the nl.tudelft.pl2016gr2.gui.view node of the root.
+   */
+  protected static ViewNode drawRootNode(IPhylogeneticTreeNode root, TreeController controller) {
+    Pane graphPane = controller.getGraphPane();
+    double startX = TreeController.GRAPH_BORDER_OFFSET;
+    double endX = graphPane.getWidth() - TreeController.GRAPH_BORDER_OFFSET;
+    double startY = TreeController.GRAPH_BORDER_OFFSET;
+    double endY = graphPane.getHeight() - TreeController.GRAPH_BORDER_OFFSET;
+    GraphArea gbox = new GraphArea(startX, endX, startY, endY);
+    return drawNode(root, null, gbox, controller);
+  }
+
+  /**
+   * Recursively draw the node and all of its children.
+   *
+   * @param dataNode the data of the node to draw.
+   * @param parent the parent of the node to draw.
+   * @param graphArea the area in which the node should be drawn.
+   * @param controller the controller of the tree.
+   * @return the drawn nl.tudelft.pl2016gr2.gui.view node.
+   */
+  private static ViewNode drawNode(IPhylogeneticTreeNode dataNode, ViewNode parent, GraphArea graphArea,
+          TreeController controller) {
+    if (graphArea.getWidth() < NODE_DIAMETER || graphArea.getHeight() < NODE_DIAMETER
+            || dataNode == null) {
+      return null; // box too small to draw node.
     }
-
-    /**
-     * Get the data of this node.
-     *
-     * @return the data of this node.
-     */
-    public IPhylogeneticTreeNode getDataNode() {
-        return dataNode;
+    ViewNode node = new ViewNode(dataNode, parent, graphArea, controller);
+    controller.getGraphPane().getChildren().add(node);
+    double nextStartX = graphArea.getCenterX();
+    double ySize = graphArea.getHeight() / dataNode.getDirectChildCount();
+    for (int i = 0; i < dataNode.getDirectChildCount(); i++) {
+      IPhylogeneticTreeNode childDataNode = dataNode.getChild(i);
+      double nextStartY = ySize * i + graphArea.startY;
+      double nextEndY = nextStartY + ySize;
+      GraphArea childArea = new GraphArea(nextStartX, graphArea.endX, nextStartY, nextEndY);
+      ViewNode child = drawNode(childDataNode, node, childArea, controller);
+      if (child == null) {
+        drawElipsis(node, controller.getGraphPane());
+        node.isLeaf = true;
+        break;
+      }
+      node.isLeaf = false;
+      node.children.add(child);
+      drawEdge(node, child, controller.getGraphPane());
     }
+    return node;
+  }
 
-    /**
-     * Draw a root node and all of its children which fit on the screen.
-     *
-     * @param root the root node.
-     * @param controller the controller of the tree.
-     * @return the nl.tudelft.pl2016gr2.gui.view node of the root.
-     */
-    protected static ViewNode drawRootNode(IPhylogeneticTreeNode root, TreeController controller) {
-        Pane graphPane = controller.getGraphPane();
-        double startX = TreeController.GRAPH_BORDER_OFFSET;
-        double endX = graphPane.getWidth() - TreeController.GRAPH_BORDER_OFFSET;
-        double startY = TreeController.GRAPH_BORDER_OFFSET;
-        double endY = graphPane.getHeight() - TreeController.GRAPH_BORDER_OFFSET;
-        GraphArea gbox = new GraphArea(startX, endX, startY, endY);
-        return drawNode(root, null, gbox, controller);
+  /**
+   * Draw the edge between a parent and a child node.
+   *
+   * @param parent the parent node.
+   * @param child the child node.
+   * @param graphPane the pane in which the edge should be drawn.
+   */
+  private static void drawEdge(ViewNode parent, ViewNode child, Pane graphPane) {
+    Line edge = new Line();
+    edge.startXProperty().bind(parent.centerXProperty());
+    edge.startYProperty().bind(parent.centerYProperty());
+    edge.endXProperty().bind(child.centerXProperty());
+    edge.endYProperty().bind(child.centerYProperty());
+    graphPane.getChildren().add(edge);
+    edge.toBack();
+    child.edge = edge;
+  }
+
+  /**
+   * Draw an elipsis after a parent node to indicate that the parent has children, but there is not
+   * enough space to draw the children.
+   *
+   * @param node the parent node.
+   * @param graphPane the pane in which to draw the elipsis.
+   */
+  private static void drawElipsis(ViewNode node, Pane graphPane) {
+    Line l = new Line();
+    l.startXProperty().bind(node.centerXProperty().add(node.getRadius() * 2));
+    l.endXProperty().bind(node.centerXProperty().add(node.getRadius() * 4));
+    l.startYProperty().bind(node.centerYProperty());
+    l.endYProperty().bind(node.centerYProperty());
+    l.getStrokeDashArray().addAll(2d, 5d);
+    graphPane.getChildren().add(l);
+    node.elipsis = l;
+  }
+
+  /**
+   * Animate the zoom in on a node.
+   *
+   * @param newRoot the node which should be the root after zooming in.
+   * @param timeline the timeline which is used for the animation.
+   */
+  public void zoomIn(ViewNode newRoot, Timeline timeline) {
+    zoomIn(this.graphArea, newRoot.graphArea, timeline);
+  }
+
+  /**
+   * Zoom in on a node and animate the zooming proces.
+   *
+   * @param originalArea the area of the currently drawn tree.
+   * @param zoomArea the area of the tree to zoom in on.
+   * @param timeline the timeline which is used for the animation.
+   */
+  private void zoomIn(GraphArea originalArea, GraphArea zoomArea, Timeline timeline) {
+    double newX = (getCenterX() - zoomArea.startX - NODE_RADIUS);
+    newX = newX / zoomArea.getWidth() * originalArea.getWidth();
+    newX += NODE_RADIUS + TreeController.GRAPH_BORDER_OFFSET;
+    double newY = (getCenterY() - zoomArea.startY);
+    newY = newY / zoomArea.getHeight() * originalArea.getHeight();
+    newY += TreeController.GRAPH_BORDER_OFFSET;
+
+    KeyValue kvX = new KeyValue(this.centerXProperty(), newX, Interpolator.EASE_BOTH);
+    KeyValue kvY = new KeyValue(this.centerYProperty(), newY, Interpolator.EASE_BOTH);
+    timeline.getKeyFrames().add(new KeyFrame(ANIMATION_DURATION, kvX, kvY));
+    fireEvent(new AnimationEvent(getCenterX(), getCenterY(), newX, newY,
+            originalArea.getHeight() / zoomArea.getHeight(), timeline, ANIMATION_DURATION));
+
+    for (ViewNode child : children) {
+      child.zoomIn(originalArea, zoomArea, timeline);
     }
+  }
 
-    /**
-     * Recursively draw the node and all of its children.
-     *
-     * @param dataNode the data of the node to draw.
-     * @param parent the parent of the node to draw.
-     * @param graphArea the area in which the node should be drawn.
-     * @param controller the controller of the tree.
-     * @return the drawn nl.tudelft.pl2016gr2.gui.view node.
-     */
-    private static ViewNode drawNode(IPhylogeneticTreeNode dataNode, ViewNode parent, GraphArea graphArea,
-            TreeController controller) {
-        if (graphArea.getWidth() < NODE_DIAMETER || graphArea.getHeight() < NODE_DIAMETER
-                || dataNode == null) {
-            return null; // box too small to draw node.
-        }
-        ViewNode node = new ViewNode(dataNode, parent, graphArea, controller);
-        controller.getGraphPane().getChildren().add(node);
-        double nextStartX = graphArea.getCenterX();
-        double ySize = graphArea.getHeight() / dataNode.getDirectChildCount();
-        for (int i = 0; i < dataNode.getDirectChildCount(); i++) {
-            IPhylogeneticTreeNode childDataNode = dataNode.getChild(i);
-            double nextStartY = ySize * i + graphArea.startY;
-            double nextEndY = nextStartY + ySize;
-            GraphArea childArea = new GraphArea(nextStartX, graphArea.endX, nextStartY, nextEndY);
-            ViewNode child = drawNode(childDataNode, node, childArea, controller);
-            if (child == null) {
-                drawElipsis(node, controller.getGraphPane());
-                node.isLeaf = true;
-                break;
-            }
-            node.isLeaf = false;
-            node.children.add(child);
-            drawEdge(node, child, controller.getGraphPane());
-        }
-        return node;
+  /**
+   * Animate the zoom out of 1 level (to the direct parent of the current root).
+   *
+   * @param timeline the timeline which is used for the animation.
+   */
+  public void zoomOut(Timeline timeline) {
+    IPhylogeneticTreeNode newRoot = dataNode.getParent();
+    double nextStartX = graphArea.getCenterX();
+    double ySize = graphArea.getHeight() / newRoot.getDirectChildCount();
+    double nextStartY = ySize * newRoot.getChildIndex(this.dataNode) + graphArea.startY;
+    double nextEndY = nextStartY + ySize;
+    GraphArea newArea = new GraphArea(nextStartX, this.graphArea.endX, nextStartY, nextEndY);
+    zoomOut(this.graphArea, newArea, timeline);
+  }
+
+  /**
+   * Zoom out and animate the zooming proces.
+   *
+   * @param originalArea the area of the currently drawn tree.
+   * @param zoomArea the area of the tree where the current tree will be drawn after zooming out.
+   * @param timeline the timeline which is used for the animation.
+   */
+  private void zoomOut(GraphArea originalArea, GraphArea zoomArea, Timeline timeline) {
+    double newX = (getCenterX() - originalArea.startX - NODE_RADIUS);
+    newX = newX * zoomArea.getWidth() / originalArea.getWidth() + zoomArea.startX;
+    newX += NODE_RADIUS;
+    double newY = (getCenterY() - originalArea.startY);
+    newY = newY * zoomArea.getHeight() / originalArea.getHeight() + zoomArea.startY;
+
+    KeyValue kvX = new KeyValue(this.centerXProperty(), newX, Interpolator.EASE_BOTH);
+    KeyValue kvY = new KeyValue(this.centerYProperty(), newY, Interpolator.EASE_BOTH);
+    timeline.getKeyFrames().add(new KeyFrame(ANIMATION_DURATION, kvX, kvY));
+    fireEvent(new AnimationEvent(getCenterX(), getCenterY(), newX, newY, 0.5, timeline,
+            ANIMATION_DURATION));
+
+    for (ViewNode child : children) {
+      child.zoomOut(originalArea, zoomArea, timeline);
     }
+  }
 
-    /**
-     * Draw the edge between a parent and a child node.
-     *
-     * @param parent the parent node.
-     * @param child the child node.
-     * @param graphPane the pane in which the edge should be drawn.
-     */
-    private static void drawEdge(ViewNode parent, ViewNode child, Pane graphPane) {
-        Line edge = new Line();
-        edge.startXProperty().bind(parent.centerXProperty());
-        edge.startYProperty().bind(parent.centerYProperty());
-        edge.endXProperty().bind(child.centerXProperty());
-        edge.endYProperty().bind(child.centerYProperty());
-        graphPane.getChildren().add(edge);
-        edge.toBack();
-        child.edge = edge;
+  public GraphArea getGraphArea() {
+    return graphArea;
+  }
+
+  public ArrayList<ViewNode> getCurrentLeaves() {
+    ArrayList<ViewNode> res = new ArrayList<>();
+    if (!isLeaf) {
+      for (ViewNode child : children) {
+        res.addAll(child.getCurrentLeaves());
+      }
+    } else {
+      res.add(this);
     }
-
-    /**
-     * Draw an elipsis after a parent node to indicate that the parent has children, but there is
-     * not enough space to draw the children.
-     *
-     * @param node the parent node.
-     * @param graphPane the pane in which to draw the elipsis.
-     */
-    private static void drawElipsis(ViewNode node, Pane graphPane) {
-        Line l = new Line();
-        l.startXProperty().bind(node.centerXProperty().add(node.getRadius() * 2));
-        l.endXProperty().bind(node.centerXProperty().add(node.getRadius() * 4));
-        l.startYProperty().bind(node.centerYProperty());
-        l.endYProperty().bind(node.centerYProperty());
-        l.getStrokeDashArray().addAll(2d, 5d);
-        graphPane.getChildren().add(l);
-        node.elipsis = l;
-    }
-
-    /**
-     * Animate the zoom in on a node.
-     *
-     * @param newRoot the node which should be the root after zooming in.
-     * @param timeline the timeline which is used for the animation.
-     */
-    public void zoomIn(ViewNode newRoot, Timeline timeline) {
-        zoomIn(this.graphArea, newRoot.graphArea, timeline);
-    }
-
-    /**
-     * Zoom in on a node and animate the zooming proces.
-     *
-     * @param originalArea the area of the currently drawn tree.
-     * @param zoomArea the area of the tree to zoom in on.
-     * @param timeline the timeline which is used for the animation.
-     */
-    private void zoomIn(GraphArea originalArea, GraphArea zoomArea, Timeline timeline) {
-        double newX = (getCenterX() - zoomArea.startX - NODE_RADIUS);
-        newX = newX / zoomArea.getWidth() * originalArea.getWidth();
-        newX += NODE_RADIUS + TreeController.GRAPH_BORDER_OFFSET;
-        double newY = (getCenterY() - zoomArea.startY);
-        newY = newY / zoomArea.getHeight() * originalArea.getHeight();
-        newY += TreeController.GRAPH_BORDER_OFFSET;
-
-        KeyValue kvX = new KeyValue(this.centerXProperty(), newX, Interpolator.EASE_BOTH);
-        KeyValue kvY = new KeyValue(this.centerYProperty(), newY, Interpolator.EASE_BOTH);
-        timeline.getKeyFrames().add(new KeyFrame(ANIMATION_DURATION, kvX, kvY));
-        fireEvent(new AnimationEvent(getCenterX(), getCenterY(), newX, newY,
-                originalArea.getHeight() / zoomArea.getHeight(), timeline, ANIMATION_DURATION));
-
-        for (ViewNode child : children) {
-            child.zoomIn(originalArea, zoomArea, timeline);
-        }
-    }
-
-    /**
-     * Animate the zoom out of 1 level (to the direct parent of the current root).
-     *
-     * @param timeline the timeline which is used for the animation.
-     */
-    public void zoomOut(Timeline timeline) {
-        IPhylogeneticTreeNode newRoot = dataNode.getParent();
-        double nextStartX = graphArea.getCenterX();
-        double ySize = graphArea.getHeight() / newRoot.getDirectChildCount();
-        double nextStartY = ySize * newRoot.getChildIndex(this.dataNode) + graphArea.startY;
-        double nextEndY = nextStartY + ySize;
-        GraphArea newArea = new GraphArea(nextStartX, this.graphArea.endX, nextStartY, nextEndY);
-        zoomOut(this.graphArea, newArea, timeline);
-    }
-
-    /**
-     * Zoom out and animate the zooming proces.
-     *
-     * @param originalArea the area of the currently drawn tree.
-     * @param zoomArea the area of the tree where the current tree will be drawn after zooming out.
-     * @param timeline the timeline which is used for the animation.
-     */
-    private void zoomOut(GraphArea originalArea, GraphArea zoomArea, Timeline timeline) {
-        double newX = (getCenterX() - originalArea.startX - NODE_RADIUS);
-        newX = newX * zoomArea.getWidth() / originalArea.getWidth() + zoomArea.startX;
-        newX += NODE_RADIUS;
-        double newY = (getCenterY() - originalArea.startY);
-        newY = newY * zoomArea.getHeight() / originalArea.getHeight() + zoomArea.startY;
-
-        KeyValue kvX = new KeyValue(this.centerXProperty(), newX, Interpolator.EASE_BOTH);
-        KeyValue kvY = new KeyValue(this.centerYProperty(), newY, Interpolator.EASE_BOTH);
-        timeline.getKeyFrames().add(new KeyFrame(ANIMATION_DURATION, kvX, kvY));
-        fireEvent(new AnimationEvent(getCenterX(), getCenterY(), newX, newY, 0.5, timeline,
-                ANIMATION_DURATION));
-
-        for (ViewNode child : children) {
-            child.zoomOut(originalArea, zoomArea, timeline);
-        }
-    }
-
-    public GraphArea getGraphArea() {
-        return graphArea;
-    }
-
-    public ArrayList<ViewNode> getCurrentLeaves() {
-        ArrayList<ViewNode> res = new ArrayList<>();
-        if (!isLeaf) {
-            for (ViewNode child : children) {
-                res.addAll(child.getCurrentLeaves());
-            }
-        } else {
-            res.add(this);
-        }
-        return res;
-    }
+    return res;
+  }
 }
