+++ PL2/PL2-gui/src/main/java/nl/tudelft/pl2016gr2/gui/view/tree/heatmap/INodeHeatmap.java
@@ -9,6 +9,6 @@ import nl.tudelft.pl2016gr2.gui.view.tree.ViewNode;
  */
 public interface INodeHeatmap {
 
-    void onChange(ArrayList<ViewNode> newLeaves);
+  void onChange(ArrayList<ViewNode> newLeaves);
 
 }
